requirements.txt
# Core ML frameworks
torch>=2.0.0
transformers>=4.35.0
accelerate>=0.24.0
bitsandbytes>=0.41.0

# Model optimization
optimum>=1.14.0
onnx>=1.15.0
onnxruntime>=1.16.0

# RAG components
langchain>=0.1.0
langchain-community>=0.0.10
chromadb>=0.4.18
sentence-transformers>=2.2.2

# API and monitoring
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6

# Utilities
numpy>=1.24.0
pandas>=2.1.0
python-dotenv>=1.0.0